Rename query loop hoc
